Allow custom session duration
Previous hardcoded session lifetime of 15m caused issues with libraries
like botocore, which have a hard set limit of refreshing token a set time
before expiry. In botocore the default happens to be 15 minutes, which
means that the session token is never cached within botocore.

This causes excess load on kiam and quite often leads to missing access
keys on the service using the tokens.

This commit adds a flag to define the lifetime of the token that is
requested from AWS API.