Replace command struct initialisation with zero-values
For all commands the defined flags should be used to change the parameters. This
change serves to make that more apparent so we're less likely to modify the
structs in `main()`.

Other than the flag bindings there is no code which modifies the TLSConfig, so
we removed that pointer and depend on the zero-struct.