Rename allowed-routes flag to whitelist-route-regexp
At the moment the flag parsing library we use will tell a user that this
parameter is a regular expression, but it's possible that a migration to cobra
or similar will make that help string disappear or change to 'string'. Since we
don't test the output of `--help`, instead make the name reflect the type.