Metadata server: Use 301 rather than 308 redirects (#173)
In 20b1c39616ce496c606912151387009692a2bdee, requests for the
security-credentials URL were changed to redirect to
security-credentials/ with a trailing slash. However, this uses
http.StatusPermanentRedirect, which corresponds to a 308, whereas
the real metadata API issues a 301 redirect
(http.StatusMovedPermanently).

Some clients do not handle the 308 correctly and fail to retrieve
credentials. We've noticed broken behaviour in the Go sdk version
1.4.10, which is used by docker-machine.