Refactor unnecessary usage of channels
Some handlers create a new local channel, send to it in the `backoff.Retry` handler, and receive from it after the `backoff.Retry` call returns with a non-nil error. This is all sequential code that needs no synchronization (and its overhead), and can be expressed without using any synchronization mechanism like channels.

This changelist removes the usage of channels where it's not required and refactors it to use local assignment.