Adds new StrictNamespaceRegex option to Server
StrictNamespaceRegexp adds start of line and end of line anchors to the
namespace regexp annotation.

This ensures that a vague namespace annotation regex such as 'red'
doesn't allow roles that contain the letters "red" anywhere in their
Role ARN to be assumed by pods.

Instead the requested role will be checked against the
regex '^red$' if strict namespace regexp is enabled.